Fix access on uninitialized data
This change initializes the variable determinant to 0. This prevents access on
a uninitialised variable.

Though I am not 100% sure, I think that the variable determinant is read but
never set when the value of log det A for a matrix A yields (numerically) 0.
From my tests this change fixes the access on uninitialized data (checked with
valgrind) and still yields correct results.